Converted indexset view tests to new file generation pattern,
Also, added unset calls to CMakeList files to avoid potential pollution across
test directories.
