+++ test/functional/forall/indexset-view/tests/test-ForallIcountIndexSetView.hpp
@@ -5,8 +5,8 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
-#ifndef __TEST_FORALL_BASIC_INDEXSET_VIEW__HPP__
-#define __TEST_FORALL_BASIC_INDEXSET_VIEW__HPP__
+#ifndef __TEST_FORALL_ICOUNT_INDEXSET_VIEW_HPP__
+#define __TEST_FORALL_ICOUNT_INDEXSET_VIEW_HPP__
 
 #include "RAJA_test-indexset-build.hpp"
 
@@ -14,8 +14,9 @@
 #include <algorithm>
 #include <vector>
 
+
 template <typename INDEX_TYPE, typename WORKING_RES, typename EXEC_POLICY>
-void ForallISetViewTest()
+void ForallIcountIndexSetViewTestImpl()
 {
 
   using RangeSegType       = RAJA::TypedRangeSegment<INDEX_TYPE>;
@@ -63,17 +64,18 @@ void ForallISetViewTest()
 
   working_res.memcpy(working_array, test_array, sizeof(INDEX_TYPE) * N);
 
+  INDEX_TYPE ticount = 0;
   for (size_t i = 0; i < is_indices.size(); ++i) {
-    test_array[ is_indices[i] ] = is_indices[i];
+    test_array[ ticount++ ] = is_indices[i];
   }
 
-  using view_type = RAJA::View< INDEX_TYPE, RAJA::Layout<1, INDEX_TYPE, 0> >;
-
   RAJA::Layout<1> layout(N);
-  view_type work_view(working_array, layout);
+  RAJA::View< INDEX_TYPE, RAJA::Layout<1, INDEX_TYPE, 0> >
+    work_view(working_array, layout);
 
-  RAJA::forall<EXEC_POLICY>(iset, [=] RAJA_HOST_DEVICE(INDEX_TYPE idx) {
-    working_array[idx] = idx;
+  RAJA::forall_Icount<EXEC_POLICY>(iset, 
+    [=] RAJA_HOST_DEVICE(INDEX_TYPE icount, INDEX_TYPE idx) {
+    work_view( icount ) = idx;
   });
 
   working_res.memcpy(check_array, working_array, sizeof(INDEX_TYPE) * N);
@@ -90,13 +92,22 @@ void ForallISetViewTest()
 }
 
 
-TYPED_TEST_P(ForallIndexSetViewTest, IndexSetForallView)
+TYPED_TEST_SUITE_P(ForallIcountIndexSetViewTest);
+template <typename T>
+class ForallIcountIndexSetViewTest : public ::testing::Test
+{
+};
+
+TYPED_TEST_P(ForallIcountIndexSetViewTest, IndexSetForallIcountView)
 {
   using INDEX_TYPE       = typename camp::at<TypeParam, camp::num<0>>::type;
   using WORKING_RESOURCE = typename camp::at<TypeParam, camp::num<1>>::type;
   using EXEC_POLICY      = typename camp::at<TypeParam, camp::num<2>>::type;
 
-  ForallISetViewTest<INDEX_TYPE, WORKING_RESOURCE, EXEC_POLICY>();
+  ForallIcountIndexSetViewTestImpl<INDEX_TYPE, WORKING_RESOURCE, EXEC_POLICY>();
 }
 
-#endif  // __TEST_FORALL_BASIC_INDEXSET_VIEW__HPP__
+REGISTER_TYPED_TEST_SUITE_P(ForallIcountIndexSetViewTest,
+                            IndexSetForallIcountView);
+
+#endif  // __TEST_FORALL_ICOUNT_INDEXSET_VIEW_HPP__
