+++ test/functional/forall/indexset-view/tests/test-ForallIndexSetView.hpp
@@ -5,8 +5,8 @@
 // SPDX-License-Identifier: (BSD-3-Clause)
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
-#ifndef __TEST_FORALL_ICOUNT_INDEXSET_VIEW__HPP__
-#define __TEST_FORALL_ICOUNT_INDEXSET_VIEW__HPP__
+#ifndef __TEST_FORALL_INDEXSET_VIEW_HPP__
+#define __TEST_FORALL_INDEXSET_VIEW_HPP__
 
 #include "RAJA_test-indexset-build.hpp"
 
@@ -14,9 +14,8 @@
 #include <algorithm>
 #include <vector>
 
-
 template <typename INDEX_TYPE, typename WORKING_RES, typename EXEC_POLICY>
-void Forall_IcountISetViewTest()
+void ForallIndexSetViewTestImpl()
 {
 
   using RangeSegType       = RAJA::TypedRangeSegment<INDEX_TYPE>;
@@ -64,18 +63,17 @@ void Forall_IcountISetViewTest()
 
   working_res.memcpy(working_array, test_array, sizeof(INDEX_TYPE) * N);
 
-  INDEX_TYPE ticount = 0;
   for (size_t i = 0; i < is_indices.size(); ++i) {
-    test_array[ ticount++ ] = is_indices[i];
+    test_array[ is_indices[i] ] = is_indices[i];
   }
 
-  RAJA::Layout<1> layout(N);
-  RAJA::View< INDEX_TYPE, RAJA::Layout<1, INDEX_TYPE, 0> >
-    work_view(working_array, layout);
+  using view_type = RAJA::View< INDEX_TYPE, RAJA::Layout<1, INDEX_TYPE, 0> >;
 
-  RAJA::forall_Icount<EXEC_POLICY>(iset, 
-    [=] RAJA_HOST_DEVICE(INDEX_TYPE icount, INDEX_TYPE idx) {
-    work_view( icount ) = idx;
+  RAJA::Layout<1> layout(N);
+  view_type work_view(working_array, layout);
+
+  RAJA::forall<EXEC_POLICY>(iset, [=] RAJA_HOST_DEVICE(INDEX_TYPE idx) {
+    working_array[idx] = idx;
   });
 
   working_res.memcpy(check_array, working_array, sizeof(INDEX_TYPE) * N);
@@ -92,13 +90,22 @@ void Forall_IcountISetViewTest()
 }
 
 
-TYPED_TEST_P(ForallIndexSetViewTest, IndexSetForall_IcountView)
+TYPED_TEST_SUITE_P(ForallIndexSetViewTest);
+template <typename T>
+class ForallIndexSetViewTest : public ::testing::Test
+{
+};
+
+TYPED_TEST_P(ForallIndexSetViewTest, IndexSetForallView)
 {
   using INDEX_TYPE       = typename camp::at<TypeParam, camp::num<0>>::type;
   using WORKING_RESOURCE = typename camp::at<TypeParam, camp::num<1>>::type;
   using EXEC_POLICY      = typename camp::at<TypeParam, camp::num<2>>::type;
 
-  Forall_IcountISetViewTest<INDEX_TYPE, WORKING_RESOURCE, EXEC_POLICY>();
+  ForallIndexSetViewTestImpl<INDEX_TYPE, WORKING_RESOURCE, EXEC_POLICY>();
 }
 
-#endif  // __TEST_FORALL_ICOUNT_INDEXSET_VIEW__HPP__
+REGISTER_TYPED_TEST_SUITE_P(ForallIndexSetViewTest,
+                            IndexSetForallView);
+
+#endif  // __TEST_FORALL_INDEXSET_VIEW_HPP__
