+++ test/functional/forall/indexset/CMakeLists.txt
@@ -27,3 +27,5 @@ foreach( BACKEND ${FORALL_BACKENDS} )
                                PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/tests)
   endforeach()
 endforeach()
+
+unset( INDEXSETTESTTYPES )
